+++ cmd/cmd.go
@@ -64,7 +64,7 @@ func configure(ctx *cli.Context) error {
 	}
 
 	if ctx.String("gitlab-admin-token") != "" {
-		c.gitlabAdmin, err = gitlab.NewClient(ctx.GlobalString("gitlab-admin-token"), opts...)
+		c.gitlabAdmin, err = gitlab.NewClient(ctx.String("gitlab-admin-token"), opts...)
 		if err != nil {
 			return err
 		}